# Conflicts:
#	app/src/main/java/com/neqabty/MainActivity.kt
#	app/src/main/java/com/neqabty/data/api/WebService.kt
#	app/src/main/java/com/neqabty/data/entities/MedicalRenewalPaymentData.kt
#	app/src/main/java/com/neqabty/data/repositories/CachedNeqabtyDataStore.kt
#	app/src/main/java/com/neqabty/domain/entities/MedicalRenewalPaymentEntity.kt
#	app/src/main/java/com/neqabty/presentation/common/Constants.kt
#	app/src/main/java/com/neqabty/presentation/di/AppModule.kt
#	app/src/main/java/com/neqabty/presentation/entities/MedicalRenewalPaymentUI.kt
#	app/src/main/java/com/neqabty/presentation/ui/activateAccount/ActivateAccountFragment.kt
#	app/src/main/java/com/neqabty/presentation/ui/inquiryDetails/InquiryDetailsFragment.kt
#	app/src/main/java/com/neqabty/presentation/ui/medicalRenewDetails/MedicalRenewDetailsFragment.kt
#	app/src/main/java/com/neqabty/presentation/ui/signup/SignupFragment.kt
#	app/src/main/res/navigation/main.xml
#	app/src/main/res/values/strings.xml
#	versions.gradle

+++ app/src/main/java/com/neqabty/data/entities/MedicalRenewalPaymentData.kt
@@ -19,7 +19,7 @@ data class MedicalRenewalPaymentData(
         @field:SerializedName("PaymentRequestNumber")
         var paymentRequestNumber: String? = "",
         @field:SerializedName("Value")
-        var amount: Int?,
+        var amount: Double?,
         @field:SerializedName("PaymentRequestName")
         var name: String?,
         @field:SerializedName("EngineerName")
